bugfixes:   * fixed issue 2: drawing nodes with many edges results in edges being shifted to the left (http://code.google.com/p/whatswrong/issues/detail?id=2)
etc:
  * added util package to this project

+++ branches/0.0.x/src/whatswrong/DependencyLayout.java
@@ -203,16 +203,15 @@ public class DependencyLayout {
       });
       //now put points along the token vertex wrt to ordering
       List<DependencyEdge> loopsOnVertex = loops.get(token);
-      int width = (int) ((tokenLayout.getBounds(token).getWidth() + vertexExtraSpace) /
-              (connections.size() + 1 + loopsOnVertex.size() * 2));
-      int x = (int) (tokenLayout.getBounds(token).getMinX() - vertexExtraSpace / 2 + width);
+      double width = (tokenLayout.getBounds(token).getWidth() + vertexExtraSpace) / (connections.size() + 1.0 + loopsOnVertex.size() * 2);
+      double x = (tokenLayout.getBounds(token).getMinX() - (vertexExtraSpace / 2.0)) + width;
       for (DependencyEdge loop : loopsOnVertex) {
-        Point point = new Point(x, baseline + maxHeight);
+        Point point = new Point((int)x, baseline + maxHeight);
         from.put(loop, point);
         x += width;
       }
       for (DependencyEdge edge : connections) {
-        Point point = new Point(x, baseline + maxHeight);
+        Point point = new Point((int)x, baseline + maxHeight);
         if (edge.getFrom().equals(token))
           from.put(edge, point);
         else
@@ -221,7 +220,7 @@ public class DependencyLayout {
 
       }
       for (DependencyEdge loop : loopsOnVertex) {
-        Point point = new Point(x, baseline + maxHeight);
+        Point point = new Point((int)x, baseline + maxHeight);
         to.put(loop, point);
         x += width;
       }
